+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## New features
+- Red majority shader effect when player has low count of lifes
+
 ## Fixes
 - All nine lifes damage types do not knockback (bleudev)
 - Use charged amethyst damage source in amethyst explosions